+++ lib/src/indicators/circular_slide_indicator.dart
@@ -1,56 +1,42 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_carousel_widget/src/indicators/models/slide_indicator_options_model.dart';
 
 import 'slide_indicator.dart';
 
 class CircularSlideIndicator implements SlideIndicator {
-  const CircularSlideIndicator({
-    this.itemSpacing = 20,
-    this.indicatorRadius = 6,
-    this.indicatorBorderWidth = 1,
-    this.indicatorBorderColor,
-    this.padding,
-    this.alignment = Alignment.bottomCenter,
-    this.currentIndicatorColor,
-    this.indicatorBackgroundColor,
-  });
+  const CircularSlideIndicator(
+      {this.slideIndicatorOptions = const SlideIndicatorOptions()});
 
-  final AlignmentGeometry alignment;
-  final Color? currentIndicatorColor;
-  final Color? indicatorBackgroundColor;
-  final Color? indicatorBorderColor;
-  final double indicatorBorderWidth;
-  final double indicatorRadius;
-  final double itemSpacing;
-  final EdgeInsets? padding;
+  final SlideIndicatorOptions slideIndicatorOptions;
 
   @override
   Widget build(int currentPage, double pageDelta, int itemCount) {
-    var activeColor = const Color(0xFFFFFFFF);
-    var backgroundColor = const Color(0x66FFFFFF);
-
-    // if (SchedulerBinding.instance.window.platformBrightness ==
-    //     Brightness.light) {
-    //   activeColor = const Color(0xFF000000);
-    //   backgroundColor = const Color.fromARGB(255, 163, 159, 159);
-    // }
-
     return Container(
-      alignment: alignment,
-      padding: padding,
-      child: SizedBox(
-        width: itemCount * itemSpacing,
-        height: indicatorRadius * 2,
-        child: CustomPaint(
-          painter: CircularIndicatorPainter(
-            currentIndicatorColor: currentIndicatorColor ?? activeColor,
-            indicatorBackgroundColor:
-                indicatorBackgroundColor ?? backgroundColor,
-            currentPage: currentPage,
-            pageDelta: pageDelta,
-            itemCount: itemCount,
-            radius: indicatorRadius,
-            indicatorBorderColor: indicatorBorderColor,
-            borderWidth: indicatorBorderWidth,
+      alignment: slideIndicatorOptions.alignment,
+      padding: slideIndicatorOptions.padding,
+      child: Container(
+        decoration: slideIndicatorOptions.enableHalo
+            ? slideIndicatorOptions.haloDecoration
+            : null,
+        padding: slideIndicatorOptions.enableHalo
+            ? slideIndicatorOptions.haloPadding
+            : null,
+        child: SizedBox(
+          width: itemCount * slideIndicatorOptions.itemSpacing,
+          height: slideIndicatorOptions.indicatorRadius * 2,
+          child: CustomPaint(
+            painter: CircularIndicatorPainter(
+              currentIndicatorColor:
+                  slideIndicatorOptions.currentIndicatorColor,
+              indicatorBackgroundColor:
+                  slideIndicatorOptions.indicatorBackgroundColor,
+              currentPage: currentPage,
+              pageDelta: pageDelta,
+              itemCount: itemCount,
+              radius: slideIndicatorOptions.indicatorRadius,
+              indicatorBorderColor: slideIndicatorOptions.indicatorBorderColor,
+              borderWidth: slideIndicatorOptions.indicatorBorderWidth,
+            ),
           ),
         ),
       ),